requirements.txt
requests==2.26.0
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.9.1+cu111
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.10.1+cu111
faiss-cpu==1.7.1.post2
transformers==4.11.3
sentence-transformers==2.1.0
python-memcached==1.59